Moved 'delete_base_command' out of 'Command's folder to hide it from the list
