+++ kpi/management/delete_base_command.py
